Day 2 Pt 1 Working
BenchmarkRunComputer-8           2357491               515 ns/op
PASS
ok      advent/day02    1.763s
