+++ shart/day01/fuel.go
