+++ shart/day01/fuelTestcases.go
@@ -1,0 +1,163 @@
+package main
+
+var testCasesGetModuleFuelReqirement = []struct {
+	description string
+	moduleMass  int
+	expected    int
+}{
+	{
+		description: "Fuel required for a 12 mass module",
+		moduleMass:  12,
+		expected:    2,
+	},
+	{
+		description: "Fuel required for a 14 mass module",
+		moduleMass:  14,
+		expected:    2,
+	},
+	{
+		description: "Fuel required for a 1969 mass module",
+		moduleMass:  1969,
+		expected:    966,
+	},
+	{
+		description: "Fuel required for a 100756 mass module",
+		moduleMass:  100756,
+		expected:    50346,
+	},
+	{
+		description: "Fuel required for a 2 mass module",
+		moduleMass:  2,
+		expected:    0,
+	},
+}
+
+var testCasesGetFuelTotal = []struct {
+	description  string
+	modulemasses []int
+	expected     int
+}{
+	{
+		description:  "Simple 1 item test",
+		modulemasses: []int{12},
+		expected:     2,
+	},
+	{
+		description:  "Two item test",
+		modulemasses: []int{12, 12},
+		expected:     4,
+	},
+	{
+		description:  "more complicated test",
+		modulemasses: []int{12, 12, 1969},
+		expected:     970,
+	},
+	{
+		description:  "Day 1 Challenge",
+		modulemasses: dayOneInput,
+		expected:     5182078,
+	},
+}
+
+var dayOneInput = []int{
+	119341,
+	141179,
+	104964,
+	90174,
+	106547,
+	78708,
+	128438,
+	84980,
+	66768,
+	106894,
+	82394,
+	95489,
+	52669,
+	95631,
+	66849,
+	135461,
+	88795,
+	77228,
+	123981,
+	72391,
+	103823,
+	63982,
+	64997,
+	145184,
+	69311,
+	117874,
+	106314,
+	101353,
+	114745,
+	88889,
+	83875,
+	85051,
+	101949,
+	118449,
+	65252,
+	135144,
+	68497,
+	132390,
+	105856,
+	121098,
+	144838,
+	87829,
+	141579,
+	140516,
+	126377,
+	55911,
+	89261,
+	76610,
+	135578,
+	110154,
+	147648,
+	106639,
+	143854,
+	91637,
+	84297,
+	117449,
+	75041,
+	58141,
+	132983,
+	114681,
+	75817,
+	116453,
+	56544,
+	65230,
+	90622,
+	133929,
+	92089,
+	134197,
+	104271,
+	127204,
+	149761,
+	128253,
+	132253,
+	109273,
+	88734,
+	136243,
+	122128,
+	119246,
+	121118,
+	128448,
+	137797,
+	66767,
+	67488,
+	90616,
+	97823,
+	51642,
+	98789,
+	94130,
+	128350,
+	101600,
+	85570,
+	145174,
+	127257,
+	141772,
+	144415,
+	82959,
+	58548,
+	129474,
+	125838,
+	68864,
+}
