+++ shart/day01/fuel_test.go
