+++ shart/go.mod
@@ -1,3 +1,5 @@
 module advent
 
 go 1.13
+
+require github.com/stretchr/testify v1.5.1
